requirements.txt
openai==1.95.0
pandas==2.3.1
promptflow==1.18.1
promptflow-tools
pyperclip==1.9.0
python-dotenv==1.1.1
Requests==2.32.4
selenium==4.34.2
streamlit==1.46.1
tqdm==4.67.1
